put mipper in own class
